Parse enough of input before CGI is forked.
We run intercept_form_submit_filter_prefetch, look for login and
password, and run pam_authenticate if needed. We store the input
in f->ctx->cached_brigade and in the intercept_form_submit_filter
we just send the cached buckets out upon the first invocation,
and passthrough during the following.