Make the comment for filter_provides_in clearer.
MUST: Packages must not provide RPM dependency information when that
information is not global in nature, or are otherwise handled (e.g.
through a virtual provides system). e.g. a plugin package containing
a binary shared library must not "provide" that library unless it is
accessible through the system library paths.

https://fedoraproject.org/wiki/Packaging:AutoProvidesAndRequiresFiltering#Summary